+++ admin/partials/tmsm-aquatonic-course-booking-admin-options-page.php
@@ -112,7 +112,8 @@
 		echo '
 	    <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
 	    <script type="text/javascript">
-	      google.charts.load("current", {"packages":["corechart"]});
+	      google.charts.load("current", {"packages":["corechart", "line"], "languages": "fr-FR"});
+	      
 	      google.charts.setOnLoadCallback(drawChart);
 	
 	      function drawChart() {
@@ -135,7 +136,7 @@
 	
 	        chart.draw(data, options);
 	      }
-	    </script><div id="piechart" style="width: 500px; height: 400px;"></div>
+	    </script><div id="piechart" style="width: 500px; height: 250px;"></div>
     ';
 
 		?>
@@ -176,7 +177,131 @@
 
 				</div>
 				</div>
+
 				</div>
+
+		<div id="dashboard-widgets" class="metabox-holder columns-1">
+			<div id="postbox-container-3" class="postbox-container">
+				<div class="meta-box-sortables ui-sortable">
+					<div class="postbox">
+						<div class="postbox-header"><h2 class="hndle ui-sortable-handle"><?php esc_html_e( 'Past Bookings by date and by type','tmsm-aquatonic-course-booking' ); ?></h2></div>
+						<div class="inside">
+							<div class="main">
+								<div id="chart_div"></div>
+								<?php
+								$bookings = new Tmsm_Aquatonic_Course_Booking_List_Table();
+								$chart_rows = '';
+
+								$bookings_total_items_by_coursestart_arrived = $bookings->get_total_past_bookings_by_coursestart_arrived();
+								$bookings_total_items_by_coursestart_cancelled = $bookings->get_total_past_bookings_by_coursestart_cancelled();
+								$bookings_total_items_by_coursestart_noshow = $bookings->get_total_past_bookings_by_coursestart_noshow();
+
+								$bookings_array = [];
+								foreach ( $bookings_total_items_by_coursestart_arrived as $item ) {
+									$bookings_array[ $item->course_start ]['arrived'] = $item->count;
+								}
+								foreach ( $bookings_total_items_by_coursestart_cancelled as $item ) {
+									$bookings_array[ $item->course_start ]['cancelled'] = $item->count;
+								}
+								foreach ( $bookings_total_items_by_coursestart_noshow as $item ) {
+									$bookings_array[ $item->course_start ]['noshow'] = $item->count;
+								}
+								ksort($bookings_array);
+								foreach ( $bookings_array as $date => $values ) {
+									$date_array = explode('-', $date);
+									$chart_rows .= '[new Date('.$date_array[0] . ', ' . intval( $date_array[1] - 1 ) .', '.intval($date_array[2]).'), '. ( intval($values['arrived']) ?? 0 ) .', '.(intval($values['cancelled']) ?? 0) .', '. (intval($values['noshow']) ?? 0 ) .'],';
+								}
+								//print_r($bookings_array);
+								?>
+								<script>
+
+                                  function drawCurveTypes() {
+                                    var data = new google.visualization.DataTable();
+                                    data.addColumn('date', '<?php echo esc_attr_e('Date', 'tmsm-aquatonic-course-booking');?>');
+                                    data.addColumn('number', '<?php echo esc_attr_e('Arrived', 'tmsm-aquatonic-course-booking');?>');
+                                    data.addColumn('number', '<?php echo esc_attr_e('Cancelled', 'tmsm-aquatonic-course-booking');?>');
+                                    data.addColumn('number', '<?php echo esc_attr_e('No-Show', 'tmsm-aquatonic-course-booking');?>');
+
+                                    data.addRows([
+										<?php echo $chart_rows ; ?>
+                                    ]);
+
+                                    var options = {
+                                      hAxis: {
+                                        title: '<?php echo esc_attr_e('Date', 'tmsm-aquatonic-course-booking');?>'
+                                      },
+                                      vAxis: {
+                                        title: '<?php echo esc_attr_e('Number', 'tmsm-aquatonic-course-booking');?>'
+                                      },
+
+                                    };
+
+                                    var chart = new google.visualization.LineChart(document.getElementById('chart_div'));
+                                    chart.draw(data, options);
+                                  }
+                                  google.charts.setOnLoadCallback(drawCurveTypes);
+
+								</script>
+							</div>
+						</div>
+					</div>
+				</div>
+			</div>
+		</div>
+
+	<div id="dashboard-widgets" class="metabox-holder columns-1">
+			<div id="postbox-container-4" class="postbox-container">
+				<div class="meta-box-sortables ui-sortable">
+					<div class="postbox" id="postbox-stats-listfuturedates">
+						<div class="postbox-header"><h2 class="hndle ui-sortable-handle"><?php esc_html_e( 'Future Bookings by date','tmsm-aquatonic-course-booking' ); ?></h2></div>
+						<div class="insidee">
+							<div class="main">
+
+								<?php
+								$bookings = new Tmsm_Aquatonic_Course_Booking_List_Table();
+								$total_future_participants_by_coursestart_active = $bookings->get_total_future_participants_by_coursestart_active();
+
+								?>
+								<div class="table-responsive">
+									<table class="wp-list-table widefat table-view-list settings_page_tmsm-aquatonic-course-booking-settings">
+										<thead>
+										<tr>
+											<th scope="col" class="manage-column column-primary"><?php esc_html_e( 'Date','tmsm-aquatonic-course-booking' ); ?></th>
+
+												<?php
+												foreach($total_future_participants_by_coursestart_active as $date){
+													$date_object = Datetime::createFromFormat( 'Y-m-d', $date->course_start);
+													$date_formatted = wp_date( 'D j M', $date_object->getTimestamp() );
+
+													echo '<td data-weekday="'.$date_object->format('N').'" '.(in_array($date_object->format('N') , [6, 7]) ? 'class="weekend alternate"' : '').'>'.$date_formatted.'</td>';
+												}
+												?>
+
+										</tr>
+										</thead>
+
+										<tbody >
+										<tr>
+											<th scope="col" class="manage-column" >
+												<?php esc_html_e( 'Future Bookings by date','tmsm-aquatonic-course-booking' ); ?>
+											</th>
+												<?php
+												foreach($total_future_participants_by_coursestart_active as $date){
+													echo '<td>'.$date->participants.'</td>';
+												}
+												?>
+
+										</tr>
+										</tbody>
+
+									</table>
+								</div>
+							</div>
+						</div>
+					</div>
+				</div>
+			</div>
+		</div>
 			</div>
 
 	<?php